requirements.txt
Pillow~=10.1.0
pypdf2~=3.0.1